Blarg

+++ core/src/com/mygdx/game/settings/GameSettings.java
@@ -3,11 +3,11 @@ package com.mygdx.game.settings;
 public class GameSettings {
 
     public int    players;
-    public double startMoney;
+    public int startMoney;
     public boolean doubleCash;
-    public double gocash;
+    public int gocash;
 
-    public GameSettings(int players,double startMoney, boolean doubleCash, double gocash){
+    public GameSettings(int players,int startMoney, boolean doubleCash, int gocash){
         this.players = players;
         this.startMoney = startMoney;
         this.doubleCash = doubleCash;
